refactor: colocate utils with table instances
