+++ src/components/InventoryViews/utils/buildGroupIdParam.ts
