+++ src/components/InventoryViews/utils/buildHostListOptions.ts
