+++ src/components/InventoryViews/utils/buildHostListParams.test.ts
@@ -1,25 +1,25 @@
 import { expect } from '@jest/globals';
 import {
-  ApiHostViewsGetHostViewsOrderByEnum,
-  ApiHostViewsGetHostViewsStalenessEnum,
-  ApiHostViewsGetHostViewsSystemTypeEnum,
-} from '@redhat-cloud-services/host-inventory-client/ApiHostViewsGetHostViews';
-import { INITIAL_INVENTORY_FILTERS } from '../DataViewFiltersContext';
-import type { InventoryFilters } from '../filters/SystemsViewFilters';
-import type { BuildHostViewsParamsInput } from './buildHostViewsParams';
-import { buildHostViewsParams } from './buildHostViewsParams';
+  ApiHostGetHostListOrderByEnum,
+  ApiHostGetHostListStalenessEnum,
+  ApiHostGetHostListSystemTypeEnum,
+} from '@redhat-cloud-services/host-inventory-client/ApiHostGetHostList';
+import { INITIAL_INVENTORY_FILTERS } from '../../SystemsView/DataViewFiltersContext';
+import type { InventoryFilters } from '../../SystemsView/filters/SystemsViewFilters';
+import type { BuildHostListParamsInput } from './buildHostListParams';
+import { buildHostListParams } from './buildHostListParams';
 import { hostQueryParamsSerializer } from './buildHostListOptions';
 
 const NOT_NIL = { is: 'not_nil' as const };
 
 const buildParams = (
-  overrides: Omit<Partial<BuildHostViewsParamsInput>, 'filters'> & {
+  overrides: Omit<Partial<BuildHostListParamsInput>, 'filters'> & {
     filters?: Partial<InventoryFilters>;
   } = {},
 ) => {
   const { filters: filterOverrides, ...rest } = overrides;
 
-  return buildHostViewsParams({
+  return buildHostListParams({
     page: 1,
     perPage: 20,
     filters: {
@@ -27,36 +27,24 @@ const buildParams = (
       ...filterOverrides,
     },
     lastSeenCustomRange: null,
-    sortBy: undefined,
-    direction: undefined,
     ...rest,
   });
 };
 
-describe('buildHostViewsParams', () => {
+describe('buildHostListParams', () => {
   describe('pagination', () => {
     it('sets page and perPage', () => {
-      const params = buildParams({ page: 3, perPage: 40 });
+      const params = buildParams({ page: 2, perPage: 50 });
 
-      expect(params.page).toBe(3);
-      expect(params.perPage).toBe(40);
+      expect(params.page).toBe(2);
+      expect(params.perPage).toBe(50);
     });
   });
 
   describe('sortBy', () => {
-    it('remaps status column sort to last_check_in', () => {
+    it('sets orderBy when sortBy is provided', () => {
       const params = buildParams({
-        sortBy: 'status' as ApiHostViewsGetHostViewsOrderByEnum,
-      });
-
-      expect(params.orderBy).toBe(
-        ApiHostViewsGetHostViewsOrderByEnum.LastCheckIn,
-      );
-    });
-
-    it('passes through API orderBy values that do not need remapping', () => {
-      const params = buildParams({
-        sortBy: ApiHostViewsGetHostViewsOrderByEnum.DisplayName,
+        sortBy: ApiHostGetHostListOrderByEnum.DisplayName,
       });
 
       expect(params.orderBy).toBe('display_name');
@@ -71,9 +59,9 @@ describe('buildHostViewsParams', () => {
 
   describe('direction', () => {
     it('sets orderHow from direction', () => {
-      const params = buildParams({ direction: 'asc' });
+      const params = buildParams({ direction: 'desc' });
 
-      expect(params.orderHow).toBe('ASC');
+      expect(params.orderHow).toBe('DESC');
     });
 
     it('omits orderHow when direction is undefined', () => {
@@ -104,18 +92,10 @@ describe('buildHostViewsParams', () => {
   describe('status', () => {
     it('sets staleness from the status filter', () => {
       const params = buildParams({
-        filters: { status: [ApiHostViewsGetHostViewsStalenessEnum.Stale] },
+        filters: { status: [ApiHostGetHostListStalenessEnum.Fresh] },
       });
 
-      expect(params.staleness).toEqual(['stale']);
-    });
-
-    it('omits staleness when the status filter is empty', () => {
-      const params = buildParams({
-        filters: { status: [] },
-      });
-
-      expect(params.staleness).toBeUndefined();
+      expect(params.staleness).toEqual(['fresh']);
     });
   });
 
@@ -127,62 +107,47 @@ describe('buildHostViewsParams', () => {
 
       expect(params.registeredWith).toEqual(['insights']);
     });
-
-    it('omits registeredWith when the source filter is empty', () => {
-      const params = buildParams({
-        filters: { source: [] },
-      });
-
-      expect(params.registeredWith).toBeUndefined();
-    });
   });
 
   describe('system_type', () => {
     it('expands image to bootc and edge', () => {
       const params = buildParams({
-        filters: { system_type: ['image'] },
+        filters: { system_type: ['image', 'conventional'] },
       });
 
       expect(params.systemType).toEqual([
-        ApiHostViewsGetHostViewsSystemTypeEnum.Bootc,
-        ApiHostViewsGetHostViewsSystemTypeEnum.Edge,
+        ApiHostGetHostListSystemTypeEnum.Bootc,
+        ApiHostGetHostListSystemTypeEnum.Edge,
+        ApiHostGetHostListSystemTypeEnum.Conventional,
       ]);
     });
+  });
 
-    it('omits systemType when the system_type filter is empty', () => {
+  describe('group_id', () => {
+    it('sets groupId and preserves empty string for ungrouped hosts', () => {
       const params = buildParams({
-        filters: { system_type: [] },
+        filters: { group_id: ['workspace-1', '', 'workspace-2'] },
       });
 
-      expect(params.systemType).toBeUndefined();
+      expect(params.groupId).toEqual(['workspace-1', 'workspace-2', '']);
+    });
+
+    it('omits groupId when group_id is empty', () => {
+      const params = buildParams({
+        filters: { group_id: [] },
+      });
+
+      expect(params.groupId).toBeUndefined();
     });
   });
 
   describe('tags', () => {
     it('sets tags from the tags filter', () => {
       const params = buildParams({
-        filters: { tags: ['env/prod'] },
+        filters: { tags: ['namespace/key=value'] },
       });
 
-      expect(params.tags).toEqual(['env/prod']);
-    });
-  });
-
-  describe('group_id', () => {
-    it('sets workspaceId and preserves empty string for ungrouped hosts', () => {
-      const params = buildParams({
-        filters: { group_id: ['workspace-1', '', 'workspace-2'] },
-      });
-
-      expect(params.workspaceId).toEqual(['workspace-1', 'workspace-2', '']);
-    });
-
-    it('omits workspaceId when group_id is empty', () => {
-      const params = buildParams({
-        filters: { group_id: [] },
-      });
-
-      expect(params.workspaceId).toBeUndefined();
+      expect(params.tags).toEqual(['namespace/key=value']);
     });
   });
 
@@ -202,7 +167,7 @@ describe('buildHostViewsParams', () => {
   });
 
   describe('options.params.fields', () => {
-    it('requests host-view system profile fields', () => {
+    it('requests host list system profile fields', () => {
       const params = buildParams();
 
       expect(params.options?.params?.fields).toEqual({
@@ -211,9 +176,6 @@ describe('buildHostViewsParams', () => {
           'system_update_method',
           'bootc_status',
           'host_type',
-          'infrastructure_type',
-          'infrastructure_vendor',
-          'workloads',
         ],
       });
     });
@@ -223,14 +185,20 @@ describe('buildHostViewsParams', () => {
     it('nests system profile filters when toolbar profile filters are set', () => {
       const params = buildParams({
         filters: {
-          workloads: ['ansible'],
+          rhcStatus: ['connected'],
+          operating_system: ['RHEL9.0'],
+          workloads: ['sap'],
         },
       });
 
       expect(params.options?.params?.filter).toEqual({
         system_profile: {
+          rhc_client_id: ['connected'],
+          operating_system: {
+            RHEL: { version: { eq: ['9.0'] } },
+          },
           workloads: {
-            ansible: NOT_NIL,
+            sap: NOT_NIL,
           },
         },
       });
@@ -251,10 +219,12 @@ describe('buildHostViewsParams', () => {
       expect(
         decodeURIComponent(
           hostQueryParamsSerializer({
-            fields: { system_profile: ['host_type'] },
+            filter: {
+              system_profile: { workloads: { sap: { is: 'not_nil' } } },
+            },
           }),
         ),
-      ).toBe('fields[system_profile][]=host_type');
+      ).toBe('filter[system_profile][workloads][sap][is]=not_nil');
     });
   });
 });
