+++ src/components/InventoryViews/utils/buildSystemProfileFilters.test.ts
