+++ src/components/InventoryViews/utils/buildSystemType.test.ts
