+++ src/components/InventoryViews/utils/buildSystemType.ts
